+++ src/biz/bokhorst/xprivacy/ActivityMain.java
@@ -80,9 +80,6 @@ public class ActivityMain extends ActivityBase implements OnItemSelectedListener
 
 	private Handler mProHandler = new Handler();
 
-	public static final String cAction = "Action";
-	public static final int cActionReboot = 1;
-
 	public static final int STATE_ATTENTION = 0;
 	public static final int STATE_CHANGED = 1;
 	public static final int STATE_SHARED = 2;
@@ -159,12 +156,6 @@ public class ActivityMain extends ActivityBase implements OnItemSelectedListener
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-
-		// Process reboot action
-		Bundle extras = getIntent().getExtras();
-		if (extras != null && extras.containsKey(cAction))
-			if (extras.getInt(cAction) == cActionReboot)
-				reboot();
 
 		// Check privacy service client
 		if (!PrivacyService.checkClient())
@@ -309,14 +300,6 @@ public class ActivityMain extends ActivityBase implements OnItemSelectedListener
 
 	@Override
 	protected void onNewIntent(Intent intent) {
-		Bundle extras = intent.getExtras();
-		if (extras != null && extras.containsKey(cAction))
-			if (extras.getInt(cAction) == cActionReboot) {
-				setIntent(intent);
-				recreate();
-				return;
-			}
-
 		if (mAppAdapter != null)
 			mAppAdapter.notifyDataSetChanged();
 	}
@@ -999,30 +982,6 @@ public class ActivityMain extends ActivityBase implements OnItemSelectedListener
 		((ScrollView) findViewById(R.id.svTutorialHeader)).setVisibility(View.VISIBLE);
 		((ScrollView) findViewById(R.id.svTutorialDetails)).setVisibility(View.VISIBLE);
 		PrivacyManager.setSetting(0, PrivacyManager.cSettingTutorialMain, Boolean.FALSE.toString());
-	}
-
-	private void reboot() {
-		AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(ActivityMain.this);
-		alertDialogBuilder.setTitle(R.string.msg_reboot);
-		alertDialogBuilder.setMessage(R.string.msg_sure);
-		alertDialogBuilder.setIcon(getThemed(R.attr.icon_launcher));
-		alertDialogBuilder.setPositiveButton(getString(android.R.string.ok), new DialogInterface.OnClickListener() {
-			@Override
-			public void onClick(DialogInterface dialog, int which) {
-				try {
-					PrivacyService.getClient().reboot();
-				} catch (Throwable ex) {
-					Util.bug(null, ex);
-				}
-			}
-		});
-		alertDialogBuilder.setNegativeButton(getString(android.R.string.cancel), new DialogInterface.OnClickListener() {
-			@Override
-			public void onClick(DialogInterface dialog, int which) {
-			}
-		});
-		AlertDialog alertDialog = alertDialogBuilder.create();
-		alertDialog.show();
 	}
 
 	// Tasks