+++ src/biz/bokhorst/xprivacy/PrivacyService.java
@@ -36,7 +36,6 @@ import android.os.Environment;
 import android.os.Handler;
 import android.os.IBinder;
 import android.os.Looper;
-import android.os.PowerManager;
 import android.os.Process;
 import android.os.RemoteException;
 import android.os.StrictMode;
@@ -1175,17 +1174,6 @@ public class PrivacyService {
 			} catch (Throwable ex) {
 				Util.bug(null, ex);
 				throw new RemoteException(ex.toString());
-			}
-		}
-
-		@Override
-		public void reboot() throws RemoteException {
-			try {
-				enforcePermission();
-				Binder.clearCallingIdentity();
-				((PowerManager) getContext().getSystemService(Context.POWER_SERVICE)).reboot(null);
-			} catch (Throwable ex) {
-				Util.bug(null, ex);
 			}
 		}